Combine UserApiService and AuctionsApiService
